btcwallet: fix non-deterministic behaviour when retrieving a custom account
In theory, it should be only one custom account with a given name. However,
users could have created custom accounts with various key scopes. In that case,
'LookupAccount' has a non deterministic behaviour. To fix that, we browse
through all key scopes (deterministacly) and return the first account we found.